fix agg test by moving check logic
